requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
idna==3.3
pycparser==2.21
pycrypto==2.6.1
pyOpenSSL==21.0.0
requests==2.27.1
six==1.16.0
termcolor==1.1.0
urllib3==1.26.8
